Scheduler
The Scheduler is the core component of MWSS. It decides automatically the best target for the next obser- vation, given the inputs described above. Both the on-sky and the simulation modes of MWSS adopt the same Scheduler code. A normal cycle of the Scheduler starts with reading the weather conditions from the ASMS-B (on-sky mode) or from the weather module (simulation mode) and the telescope conditions from the OCS (on-sky mode) or the telescope module (simulation mode) at the present time of the night. If both the weather and telescope conditions are good for observations, the Scheduler will select all observable fields and assign rank values to them. The candidate field with the best rank is then adopted as the next target. The Scheduler then move to the next cycle for the next visit